backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0
python-dotenv==1.0.1
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings>=2.0.0
boto3==1.35.24
redis==5.0.8
celery==5.4.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
replicate>=0.25.0
slowapi==0.1.9
numpy>=1.24.0
requests>=2.31.0
ultralytics>=8.3.0
stripe>=8.0.0,<10.0.0
